requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.8.2
ccxt==4.4.33
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
packaging==24.2
propcache==0.2.0
pycares==4.4.0
pycparser==2.22
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
python-dotenv==1.0.0
